新增halfMoveClock作為遊戲結束判定的條件

+++ Game/GameBoard.h
@@ -33,7 +33,7 @@ protected:
 
 	// How many moves both players have made since the last pawn advance or piece capture.
 	// Important for determining of draw.
-	size_t halfmoveClock = 0;
+	int halfmoveClock = 0;
 
 	Team currentPlayer;
 public:
